Adding remove()
I am having issues with my IDE I think, issues like the else statement at the end printing regardless of commenting it out. I had to restart my IDE to get it to work. So right now it crashes on the [] overload even though its worked this whole time. My apologies if its just my code that's causing the crashes on my end. I still need to add in code to update the data for removal of items, just trying to get it to function properly first.